Float

+++ app/Utils/Keuangan.php
@@ -119,11 +119,11 @@ class Keuangan
         $saldo_kredit = 0;
         foreach ($rek->kom_saldo as $kom_saldo) {
             if ($kom_saldo->bulan == 0) {
-                $awal_debit += intval($kom_saldo->debit);
-                $awal_kredit += intval($kom_saldo->kredit);
+                $awal_debit += floatval($kom_saldo->debit);
+                $awal_kredit += floatval($kom_saldo->kredit);
             } else {
-                $saldo_debit += intval($kom_saldo->debit);
-                $saldo_kredit += intval($kom_saldo->kredit);
+                $saldo_debit += floatval($kom_saldo->debit);
+                $saldo_kredit += floatval($kom_saldo->kredit);
             }
         }
